+++ changelog.md
@@ -5,8 +5,16 @@
 ## Unreleased
 
 ### Add
+- Save settings
+- Quality video option
+- No subtitles option
+- Clean cache
 - Search for tv shows
 - Search for subtitles tv show
+- Cache subtitles
+
+## [1.0.3] - 2016-05-19
+- Save subtitles in `~/.moov/subs`
 
 ---